+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/android/_internal/calculateAndroidNamespace.kt
@@ -1,12 +1,34 @@
 package com.javiersc.hubdle.extensions.kotlin.android._internal
 
+import com.android.build.api.dsl.ApplicationExtension
+import com.android.build.api.dsl.LibraryExtension
 import com.javiersc.gradle.properties.extensions.getBooleanProperty
 import com.javiersc.hubdle.HubdleProperty.Android
+import com.javiersc.hubdle.extensions.shared.features.hubdleJavaVersionFeature
 import com.javiersc.kotlin.stdlib.remove
 import java.io.File
 import org.gradle.api.Project
+import org.gradle.kotlin.dsl.configure
 import org.gradle.kotlin.dsl.findByType
 import org.jetbrains.kotlin.gradle.dsl.KotlinProjectExtension
+
+internal fun Project.configureAndroidApplicationJavaVersion() {
+    configure<ApplicationExtension> {
+        compileOptions {
+            sourceCompatibility = hubdleJavaVersionFeature.jvmVersion.get()
+            targetCompatibility = hubdleJavaVersionFeature.jvmVersion.get()
+        }
+    }
+}
+
+internal fun Project.configureAndroidLibraryJavaVersion() {
+    configure<LibraryExtension> {
+        compileOptions {
+            sourceCompatibility = hubdleJavaVersionFeature.jvmVersion.get()
+            targetCompatibility = hubdleJavaVersionFeature.jvmVersion.get()
+        }
+    }
+}
 
 internal fun Project.calculateAndroidNamespace(initial: String?): String? =
     when {