+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/android/features/HubdleKotlinAndroidBuildFeaturesExtension.kt
@@ -10,6 +10,7 @@ import com.javiersc.hubdle.extensions.apis.HubdleConfigurableExtension
 import com.javiersc.hubdle.extensions.apis.HubdleEnableableExtension
 import com.javiersc.hubdle.extensions.kotlin.android.application.hubdleAndroidApplication
 import com.javiersc.hubdle.extensions.kotlin.android.library.hubdleAndroidLibrary
+import com.javiersc.hubdle.extensions.kotlin.multiplatform.targets.hubdleKotlinMultiplatformAndroid
 import javax.inject.Inject
 import org.gradle.api.Project
 import org.gradle.api.provider.Property
@@ -24,7 +25,12 @@ constructor(
     override val isEnabled: Property<Boolean> = property { true }
 
     override val oneOfExtensions: Set<HubdleEnableableExtension>
-        get() = setOf(hubdleAndroidApplication, hubdleAndroidLibrary)
+        get() =
+            setOf(
+                hubdleAndroidApplication,
+                hubdleAndroidLibrary,
+                hubdleKotlinMultiplatformAndroid,
+            )
 
     override val priority: Priority = Priority.P4